Add automatic service data handling
This adds the ability to put all service data in consistently named
top level hiera values, with both individual values and arrays
of data available for use. Naming is similar to facter, but as this
is a part of the hierarchy it can be overridden using other data.

Naming is as follows:

For each service, there will be one entry per property, looking like:
"#{ServiceID}_#{property}" giving the value from the first node that
has registered that service.

There is also an array of all properties in a list:
"#{ServiceID}_#{property}_array"

Finally, there is also one entry per property per node, to access a
particular node's data:
"#{ServiceID}_#{property}_#{Node}"

Where 'ServiceID' and 'Node' are from the consul HTTP API.

This feature can be used by merely putting 'services' as a path
in the consul hierarchy.